requirements.txt
Flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-bootstrap
python-jenkins
werkzeug
rq